[general] - refactor mock data to be seperated into multiple files
