+++ src/data/promotions.ts
@@ -1,0 +1,6 @@
+import type { Promotion } from '../types';
+
+export const MOCK_PROMOTIONS: Promotion[] = [
+    { id: 'promo-1', title: 'Weekend Special', description: 'Earn 2x points on all food purchases this weekend.' },
+    { id: 'promo-2', title: 'Happy Hour Drinks', description: 'Get 50% off all drinks from 4 PM to 6 PM on weekdays.' },
+];
