+++ src/data/rewardTiers.ts
@@ -1,0 +1,8 @@
+import type { RewardTier } from '../types';
+
+export const MOCK_REWARD_TIERS: RewardTier[] = [
+    { name: 'Bronze', points: 0 },
+    { name: 'Silver', points: 1000 },
+    { name: 'Gold', points: 2500 },
+    { name: 'Platinum', points: 5000 },
+];
