+++ src/data/user.ts
@@ -1,0 +1,10 @@
+import type { User } from '../types';
+
+export const MOCK_USER: User = {
+  id: 'user-john-doe',
+  name: 'John Doe',
+  points: 1250,
+  currentTier: 'Silver',
+  nextTier: 'Gold',
+  pointsToNextTier: 2500,
+};
